Reverse rotations
Reverse the rotations; swap the positive and negative rotation calculations for
each axis. Rotation about an axis in three-dimensional space typically (and
mathematically for that matter) occurs in a counter-clockwise fashion around
said axis, assuming the angle of rotation is positive (the rotation runs
clockwise when the angle is negative). A similar thing is achieved here, as the
elements are now exchanged in a counter-clockwise order during rotation.