Modify state via keyboard input
Set the terminal to noncanonical mode, allowing keyboard input to be read
immediately by the program and processed accordingly.

The third-order tensor can be rotated:
* about the x-axis (w and s keys)
* about the y-axis (a and d keys)
* about the z-axis (q and e keys)

Additionally, since only the front section (the section furthest along the
negative z-axis, closest to the viewer) of the third-order tensor is displayed,
the up and down arrow keys can be used to change which section to display.